X86 next

+++ arch/x86/cpu/x86_64/include/arch_types.h
@@ -24,7 +24,7 @@
 #define _ARCH_TYPES_H__
 
 /** cpu specific types */
-typedef unsigned int irq_flags_t;
+typedef unsigned long irq_flags_t;
 typedef unsigned long virtual_addr_t; /* 64-bit virtual addresses */
 typedef unsigned long virtual_size_t;
 typedef unsigned long physical_addr_t;
